Update to the Date compare in the exif data test
I was encountering a failure on the exif data test on my Mac and Debian installs of the package. Changing to have both the original and the comparison on GMT solves the issue of running on PDT. Tested on node 4.1.1, 4.2.3, and 6.9.4 on Mac OSX 10.12.2, and Debian Wheezy amd64 running on a AWS EC2 hvm, with node 4.7.2.